% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Lumos app server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{shiny inputs}

\item{output}{shiny outputs}

\item{session}{shiny session}
}
\value{
a shiny app server
}
\description{
Lumos app server
}
