% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_aics.R
\name{arma_aics}
\alias{arma_aics}
\title{ARIMA Model Fit}
\usage{
arma_aics(x, P, d, Q)
}
\arguments{
\item{x}{time series input, AR, I , MA}
}
\value{
Returns the dataframe having P(AR) as row co-ordinates and Q(MA) as column co-ordinates
}
\description{
Fitting models with different AR,MA process
}
