% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\docType{data}
\name{auto}
\alias{auto}
\title{Auto data}
\format{An object of class \code{data.frame} with 397 rows and 9 columns.}
\usage{
data(auto)
}
\description{
Part of the cars data set
}
\examples{
data(auto)



}
\keyword{datasets}
