% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_lux}
\alias{calculate_lux}
\title{Calculate lux}
\usage{
calculate_lux(lumens, distance)
}
\arguments{
\item{lumens}{the amount of light being emitted from the source in lumens}

\item{distance}{the distance to the light source in cm}
}
\value{
a number indicating the lux that the light source gives
}
\description{
Given an amount of light emitted from a source in lumens and a distance to
the light source, calculate the level of light in lux
}
