% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candelabra.R
\name{candelabra}
\alias{candelabra}
\title{candelabra class}
\value{
an object with class 'candelabra'
}
\description{
candelabra class

candelabra class
}
\details{
Each candelabra is made up of 1 or more candles
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{x_position}}{the x coordinate (in cm) of the centre of the candelabra
on the stage}

\item{\code{y_position}}{the y coordinate (in cm) of the centre of the candelabra
on the stage}

\item{\code{n_candles}}{the number of the candles in the candelabra}

\item{\code{radius}}{the radius (in cm) of the candelabra}

\item{\code{height}}{the height (in cm) of the base of the candelabra from the
floor of the stage}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{id}}{returns the generated unique ID of the candelabra}

\item{\code{stage}}{returns the name of the stage on which this candelabra sits}

\item{\code{candles}}{returns a list of the candle objects contained in this candelabra}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-candelabra-new}{\code{candelabra$new()}}
\item \href{#method-candelabra-move_to}{\code{candelabra$move_to()}}
\item \href{#method-candelabra-clone}{\code{candelabra$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-candelabra-new"></a>}}
\if{latex}{\out{\hypertarget{method-candelabra-new}{}}}
\subsection{Method \code{new()}}{
create a new object of class 'candelabra'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{candelabra$new(
  stage,
  centre_position,
  n_candles,
  radius,
  candle_lumens,
  candle_height,
  height
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stage}}{the object with class 'stage' to which this candelabra belongs}

\item{\code{centre_position}}{a numeric vector of length 2 giving the (x,y) coordinates
of this candelabra on the stage}

\item{\code{n_candles}}{the number of candle objects to initiate in the candelabra}

\item{\code{radius}}{the radius of the candelabra in cm}

\item{\code{candle_lumens}}{the amount of light each candle gives off in lumens}

\item{\code{candle_height}}{the height of each candle in cm}

\item{\code{height}}{the height of the candelabra from the floor in cm}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an object of class candelabra containing multiple objects of class candle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-candelabra-move_to"></a>}}
\if{latex}{\out{\hypertarget{method-candelabra-move_to}{}}}
\subsection{Method \code{move_to()}}{
move the candelabra to a new position on the stage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{candelabra$move_to(x, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{the new x coordinate of the candelabra}

\item{\code{y}}{the new y coordinate of the candelabra}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
updates the x_position and y_position fields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-candelabra-clone"></a>}}
\if{latex}{\out{\hypertarget{method-candelabra-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{candelabra$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
