% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_prices.R
\docType{data}
\name{clock_prices}
\alias{clock_prices}
\title{Clock Prices data}
\format{An object of class \code{data.frame} with 32 rows and 3 columns.}
\usage{
data(clock_prices)
}
\description{
Clock price have 1 dependent variable price and 2 independent variable price and number of bidders.
}
\examples{
data(clock_prices)



}
\keyword{datasets}
