% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default}
\alias{default}
\title{Get lumos defaults}
\usage{
default(what)
}
\arguments{
\item{what}{the name of the default value to get}
}
\value{
a default numeric value for 'what'
}
\description{
Get lumos defaults
}
