% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumos.R
\name{lumos_on}
\alias{lumos_on}
\alias{lumos_off}
\title{Turn \code{lumos} On or Off}
\usage{
lumos_on()

lumos_off()
}
\value{
\code{NULL} Called for its side effects.
}
\description{
\code{lumos} can be turned on or off using an option. These are convenience
functions for doing so. When \code{lumos} is off, calling it will do
nothing; this can be useful to save time when running scripts in batch mode
where this output is not needed.
}
