% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R
\name{normtest}
\alias{normtest}
\title{Normtest}
\usage{
normtest(x)
}
\arguments{
\item{x}{list of numbers to be tested for normality}
}
\value{
Returns the dataframe having shapiro test, ad test, cvm test, ll test, sf test
}
\description{
Performing Test for normality
}
