% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{sld}
\alias{sld}
\title{Calculate the straight line distance between two points}
\usage{
sld(x1, y1, z1, x2, y2, z2)
}
\arguments{
\item{x1}{the x coordinate of the first point}

\item{y1}{the y coordinate of the first point}

\item{z1}{the z coordinate of the first point}

\item{x2}{the x coordinate of the second point}

\item{y2}{the y coordinate of the second point}

\item{z2}{the z coordinate of the second point}
}
\value{
a single number indicating the straight line distance
between the two points
}
\description{
Calculate the straight line distance between two points
}
