% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_spinner}
\alias{with_spinner}
\title{Add shinycssloader to a UI element}
\usage{
with_spinner(elem)
}
\arguments{
\item{elem}{the ui element to add the spinner to}
}
\value{
the input element with a loading spinner when
computing
}
\description{
Add shinycssloader to a UI element
}
