% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumos.R
\name{lumos}
\alias{lumos}
\alias{l}
\title{Shed Light on Your Data}
\usage{
lumos(
  data = NULL,
  ...,
  .drop = TRUE,
  .max = 20,
  .pct = TRUE,
  .order.by.freq = .pct,
  .blanks = TRUE,
  .gen = FALSE,
  .kable = TRUE
)

l(
  data = NULL,
  ...,
  .drop = TRUE,
  .max = 20,
  .pct = TRUE,
  .order.by.freq = .pct,
  .blanks = TRUE,
  .gen = FALSE,
  .kable = TRUE
)
}
\arguments{
\item{data}{A \code{data.frame} (or, and object that can be converted to one
by call \code{link{as.data.frame}}).}

\item{...}{Further argument, evaluated within \code{data} (so they can
directly refer to columns in data without quotes).}

\item{.drop}{If \code{TRUE}, unused factor levels are dropped.}

\item{.max}{For a single categorical variable, the maximum number of unique
categories to show (can be \code{Inf}).  For a single numberic variable the
maximum number of unique values for a numeric variable to be treated as
categorical.}

\item{.pct}{If \code{TRUE}, show percents along with counts for single
categorical variables.}

\item{.order.by.freq}{If \code{TRUE}, for a single categorical variable,
show the categories in decreasing order of frequency, from top to bottom
(i.e. show the most frequent categories on top).}

\item{.blanks}{If \code{TRUE}, insert blank spaces instead of repeating
consecutive values that are identical.}

\item{.gen}{If \code{TRUE}, instead of the usual output, run a "code
generation" procedure and print its output, then return \code{NULL}
invisibly. In this case \code{...} is ignored.}

\item{.kable}{If \code{TRUE}, call \code{\link[knitr]{kable}} on the final
object before returning.}
}
\value{
The value returned depends on the parameters. If \code{.kable} is
\code{TRUE} (the default) then an object of class \code{knitr_kable},
otherwise a \code{data.frame}. See Examples.
}
\description{
Like a magic wand that allows you to explore a \code{data.frame}.  It is so
powerful, it can be abbreviated to the single letter \code{l}.
}
\examples{
lumos(iris)
lumos(iris, .gen=TRUE)  # Generate code statements to call lumos() on each column of iris.
lumos(iris, Species)

lumos(mtcars)
lumos(mtcars, wt)
lumos(mtcars, cyl)
lumos(mtcars, cyl, .pct=FALSE)
lumos(mtcars, cyl, gear)
lumos(mtcars, cyl, gear, am)
lumos(mtcars, cyl, gear, am, .blanks=FALSE, .kable=FALSE)
}
