% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage.R
\name{stage}
\alias{stage}
\title{candelabra class}
\value{
an object with class 'candelabra'
}
\description{
candelabra class

candelabra class
}
\details{
Each candelabra is made up of 1 or more candles
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{an alias (human-readable) name for the stage}

\item{\code{height}}{the default height of candelabras on the stage (in cm)}

\item{\code{width}}{the width of the stage (from wing to wing) in cm}

\item{\code{depth}}{the depth of the stage (front to back) in cm}

\item{\code{unit_scaling}}{the size (in cm) of each stage unit (which will have
lux individually calculated). Unit scaling of 1 means each 1cm square will
have its lux calculated (lower number will result in larger computation times)}

\item{\code{max_lux}}{a fixed comparison level of lux to ensure plots are always comparable to}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{id}}{a unique generated ID for the stage}

\item{\code{alias}}{the name of the stage if it has one, otherwise the unique ID}

\item{\code{candelabras}}{a list of the candelabras currently on the stage}

\item{\code{candle_positions}}{a tibble giving the position of each candle currently
on the stage along with the candelabra it sits in and the lumens it is giving off}

\item{\code{lux_grid}}{a tibble giving the amount of lux at each stage unit and a
label indicating how much each candelabra is contributing to the total. The key
tibble used for plotting}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-stage-new}{\code{stage$new()}}
\item \href{#method-stage-reactive}{\code{stage$reactive()}}
\item \href{#method-stage-invalidate}{\code{stage$invalidate()}}
\item \href{#method-stage-change_name}{\code{stage$change_name()}}
\item \href{#method-stage-add_candelabra}{\code{stage$add_candelabra()}}
\item \href{#method-stage-move_candelabra}{\code{stage$move_candelabra()}}
\item \href{#method-stage-remove_candelabra}{\code{stage$remove_candelabra()}}
\item \href{#method-stage-plot}{\code{stage$plot()}}
\item \href{#method-stage-plot_light}{\code{stage$plot_light()}}
\item \href{#method-stage-clone}{\code{stage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stage-new"></a>}}
\if{latex}{\out{\hypertarget{method-stage-new}{}}}
\subsection{Method \code{new()}}{
create a new stage object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stage$new(
  n_candelabras = 0,
  width = 1200,
  depth = 800,
  name = "new_stage",
  max_lux = 10,
  unit_scaling = default("unit_scaling")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_candelabras}}{the number of candelabras to initiate on the stage}

\item{\code{width}}{the width of the stage in cm}

\item{\code{depth}}{the depth of the stage in cm}

\item{\code{name}}{a human-readable alias for the stage}

\item{\code{max_lux}}{a comparison level of lux to always be able to compare to}

\item{\code{unit_scaling}}{the square size in cm of each stage unit}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stage-reactive"></a>}}
\if{latex}{\out{\hypertarget{method-stage-reactive}{}}}
\subsection{Method \code{reactive()}}{
return a reactive version of the object that can be used in
a shiny app
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stage$reactive()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stage-invalidate"></a>}}
\if{latex}{\out{\hypertarget{method-stage-invalidate}{}}}
\subsection{Method \code{invalidate()}}{
invalidate the object, in a shiny setting
This should be run whenever something happens that the reactive shiny
context needs to be aware of
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stage$invalidate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stage-change_name"></a>}}
\if{latex}{\out{\hypertarget{method-stage-change_name}{}}}
\subsection{Method \code{change_name()}}{
change the name of the stage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stage$change_name(new_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_name}}{name to change the stage name to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stage-add_candelabra"></a>}}
\if{latex}{\out{\hypertarget{method-stage-add_candelabra}{}}}
\subsection{Method \code{add_candelabra()}}{
add a candelabra to the stage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stage$add_candelabra(x, y, n = 10, r = 30, l = 10, ch = 15, h = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{the x coordinate of the centre of the new candelabra}

\item{\code{y}}{the y coordinate of the centre of the new candelabra}

\item{\code{n}}{the number of candles in the new candelabra}

\item{\code{r}}{the radius in cm of the new candelabra}

\item{\code{l}}{the lumens of each candle in the new candelabra}

\item{\code{ch}}{the height of each candle in the candelabra in cm}

\item{\code{h}}{the height of the new candelabra}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
adds a new candelabra object to the stage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stage-move_candelabra"></a>}}
\if{latex}{\out{\hypertarget{method-stage-move_candelabra}{}}}
\subsection{Method \code{move_candelabra()}}{
move a candelabra on the stage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stage$move_candelabra(i, new_x, new_y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{the index of the candelabra}

\item{\code{new_x}}{the new x coordinate of the candelabra}

\item{\code{new_y}}{the new y coordinate of the candelabra}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stage-remove_candelabra"></a>}}
\if{latex}{\out{\hypertarget{method-stage-remove_candelabra}{}}}
\subsection{Method \code{remove_candelabra()}}{
remove a candelabra from the stage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stage$remove_candelabra(elem)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{elem}}{the number of the candelabra (i.e. 1 is the first candelabra
that was placed on the stage, 2 the second and so on)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
removes the specified candelabra from the stage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stage-plot"></a>}}
\if{latex}{\out{\hypertarget{method-stage-plot}{}}}
\subsection{Method \code{plot()}}{
plot candelabra positions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stage$plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a plotly plot of the stage with the candelabra and candle positions marked
on it
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stage-plot_light"></a>}}
\if{latex}{\out{\hypertarget{method-stage-plot_light}{}}}
\subsection{Method \code{plot_light()}}{
plot the stage lighting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stage$plot_light(show_candelabras = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{show_candelabras}}{logical, if TRUE the candelabra positions will
be shown on the lux plot. If FALSE they will not be visible (but the light
they produce still will be)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a plotly plot of the stage with each stage unit showing how much lux
it has
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-stage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
